Converted Types to Double

+++ include/matrix.h
@@ -8,19 +8,19 @@ class Matrix {
 public:
     int rows;
     int columns;
-    std::vector<std::vector<int>> values;
+    std::vector<std::vector<double>> values;
 
     // Constructors
-    Matrix(int rs, int cs, std::vector<std::vector<int>> &vals);
+    Matrix(int rs, int cs, std::vector<std::vector<double>> &vals);
     Matrix(int rs, int cs);
 
     // Core Functions
     void generateRandomValues();
-    Matrix add(int lambda, const Matrix &M);
-    Matrix sMultiply(int lambda);
+    Matrix add(double lambda, const Matrix &M);
+    Matrix sMultiply(double lambda);
     Matrix multiply(const Matrix &M);
     Matrix transpose();
-    std::tuple<std::vector<int>, char, int, int> findLeast();
+    std::tuple<std::vector<double>, char, int, int> findLeast();
     int calcDet();
     Matrix removeij(int ri, int rj);
